Refactoring to SwiftData
