+++ PokeDex/Pokemon.swift
@@ -1,24 +1,34 @@
 //
-//  FetchedPokemon.swift
+//  Pokemon.swift
 //  PokeDex
 //
 //  Created by Aakash Ambodkar
 //
+//
 
 import Foundation
+import SwiftData
+import SwiftUI
 
-struct FetchedPokemon: Decodable {
-    let id: Int16
-    let name: String
-    let types: [String]
-    let hp: Int16
-    let attack: Int16
-    let defense: Int16
-    let SpecialAttack: Int16
-    let SpecialDefense: Int16
-    let speed: Int16
-    let spriteURL: URL
-    let shinyURL: URL
+
+@Model
+class Pokemon: Decodable {
+    @Attribute(.unique)var id: Int
+    var name: String
+    var types: [String]
+    var hp: Int
+    
+    var attack: Int
+    var defense: Int
+    var specialAttack: Int
+    var specialDefense: Int
+    var speed: Int
+    
+    var spriteURL: URL
+    var shinyURL: URL
+    var sprite: Data?
+    var shiny: Data?
+    var favorite: Bool = false
     
     enum CodingKeys: CodingKey {
         case id
@@ -46,10 +56,10 @@ struct FetchedPokemon: Decodable {
         }
     }
     
-    init(from decoder: any Decoder) throws {
+    required init(from decoder: any Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         
-        id = try container.decode(Int16.self, forKey: .id)
+        id = try container.decode(Int.self, forKey: .id)
         
         name = try container.decode(String.self, forKey: .name)
         
@@ -70,24 +80,87 @@ struct FetchedPokemon: Decodable {
         
         types = decodedTypes
         
-        var decodedStats: [Int16] = []
+        var decodedStats: [Int] = []
         var statsContainer = try container.nestedUnkeyedContainer(forKey: .stats)
         while !statsContainer.isAtEnd {
             let statsDcitionaryContainer = try statsContainer.nestedContainer(keyedBy: CodingKeys.StatDictionaryKeys.self)
-            let stat = try statsDcitionaryContainer.decode(Int16.self, forKey: .baseStat)
+            let stat = try statsDcitionaryContainer.decode(Int.self, forKey: .baseStat)
             decodedStats.append(stat)
         }
         
         hp = decodedStats[0]
         attack = decodedStats[1]
         defense = decodedStats[2]
-        SpecialAttack = decodedStats[3]
-        SpecialDefense = decodedStats[4]
+        specialAttack = decodedStats[3]
+        specialDefense = decodedStats[4]
         speed = decodedStats[5]
         
         let spriteContainer = try container.nestedContainer(keyedBy: CodingKeys.SpriteKeys.self, forKey: .sprites)
         spriteURL = try spriteContainer.decode(URL.self, forKey: .spriteURL)
         shinyURL = try spriteContainer.decode(URL.self, forKey: .shinyURL)
+    }
+    
+    var spriteImage: Image {
+        if let data = sprite, let image = UIImage(data: data) {
+            Image(uiImage: image)
+        } else {
+            Image(.bulbasaur)
+        }
+    }
+    
+    var shinyImage: Image {
+        if let data = shiny, let image = UIImage(data: data) {
+            Image(uiImage: image)
+        } else {
+            Image(.shinybulbasaur)
+        }
+    }
+    
+    var background: ImageResource {
+        switch types[0] {
+        case "rock", "ground", "steel", "fighting", "ghost", "dark", "psychic":
+                .rockgroundsteelfightingghostdarkpsychic
+        
+        case "fire", "dragon":
+                .firedragon
+        
+        case "flying", "bug":
+                .flyingbug
+        
+        case "ice":
+                .ice
+        
+        case "water":
+                .water
+        
+        default:
+                .normalgrasselectricpoisonfairy
+        }
+    }
+    
+    var typeColor: Color {
+        Color(types[0].capitalized)
+    }
+    
+    var stats: [Stat] {
+        [
+            Stat(id: 1, name: "HP", value: hp),
+            Stat(id: 2, name: "Attack", value: attack),
+            Stat(id: 3, name: "Defense", value: defense),
+            Stat(id: 4, name: "Special Attack", value: specialAttack),
+            Stat(id: 5, name: "Sepcial Defense", value: specialDefense),
+            Stat(id: 6, name: "Speed", value: speed)
+        ]
+    }
+    
+    var highestStat: Stat {
+        stats.max { $0.value < $1.value }!
+    }
+    
+    struct Stat: Identifiable {
+        let id: Int
+        let name: String
+        let value: Int
     }
 }
